requirements.txt
httpx==0.24.1
pillow==10.0.0
playwright==1.36.0
asyncio==3.4.3